+++ VisualChatGPTStudioShared/ToolWindows/CodeReview/TerminalWindowCodeReviewControl.cs
@@ -65,7 +65,7 @@ namespace JeffPires.VisualChatGPTStudio.ToolWindows
             {
                 EnableDisableButtons(true);
 
-                CodeReviews = new List<CodeReviewItem>();
+                CodeReviews = [];
 
                 Patch changes = GitChanges.GetCurrentChanges();
 
@@ -242,7 +242,7 @@ namespace JeffPires.VisualChatGPTStudio.ToolWindows
         /// <returns>A task that represents the asynchronous operation, resulting in a CodeReviewItem containing the review details.</returns>
         private async Task<CodeReviewItem> DoCodeReview(PatchEntryChanges change)
         {
-            string codeReview = await ChatGPT.GetResponseAsync(options, options.CodeReviewCommand, change.Patch, options.StopSequences.Split(','), cancellationTokenSource.Token);
+            string codeReview = await ChatGPT.GetResponseAsync(options, options.CodeReviewCommand, change.Patch, options.StopSequences.Split([','], StringSplitOptions.RemoveEmptyEntries), cancellationTokenSource.Token);
 
             GitChanges.SeparateCodeChanges(change.Patch, out string originalCode, out string alteredCode);