+++ ReleaseNotes.md
@@ -1,5 +1,12 @@
 ﻿## Release Notes 📜
 
+### 3.2.0
+
+- Added the possibility to use completion API instead of Chat API. Thanks [karburator](https://github.com/karburator) for the collaboration.
+- Fixed the issue preventing the extension from loading correctly starting from version 17.12.0 of Visual Studio 2022.
+- Fixed the 'Stream was not readable' bug.
+- Fixed an issue where an empty value was sent in the Stop Sequences parameter when it was not defined in the extension options. This caused some models to not respond correctly.
+
 ### 3.1.0
 
 - Added the possibility to use your Microsoft Account to authenticate to the Azure Open AI Service through Entra ID.